refactor(trace): modularize trace module
