+++ crates/ethernity-deeptrace/src/trace/mod.rs
@@ -1,0 +1,7 @@
+mod detector;
+mod tree;
+mod types;
+
+pub use detector::TraceDetector;
+pub use tree::{CallNode, CallTree};
+pub use types::{CallTrace, CallType};
